+++ src/proxy-api/steps/getExcerptAndHeaderImage.ts
@@ -16,12 +16,14 @@ export default (config: ConfigService, confluence: ConfluenceService): Step => a
   if (blogImgSrc && !blogImgSrc.startsWith('http')) {
     // not a URL (image uploaded to Confluence)
     const attachments = await confluence.getAttachments(context.getPageId());
-    const blogImgAttachment = attachments.find((e) =>
-    // find the attachment matching the UID got from the headerImage attribute
-      e.fileId === blogImgSrc);
-    if (blogImgAttachment) {
-      blogImgSrc = `${webBasePath}/wiki${blogImgAttachment?.downloadLink}`;
-      context.setHeaderImage(blogImgSrc);
+    if (attachments) {
+      const blogImgAttachment = attachments.find((e) =>
+      // find the attachment matching the UID got from the headerImage attribute
+        e.fileId === blogImgSrc);
+      if (blogImgAttachment) {
+        blogImgSrc = `${webBasePath}/wiki${blogImgAttachment?.downloadLink}`;
+        context.setHeaderImage(blogImgSrc);
+      }
     }
   }
 
@@ -39,20 +41,20 @@ export default (config: ConfigService, confluence: ConfluenceService): Step => a
   // defined in a page-properties section in a blog post
   // a macro page-properties with an image and blockquote inside will be used alternatively to define
   // both image and blockquote for the blog post
-  $(".plugin-tabmeta-details[data-macro-name='details']")
-    .first()
-    .each((_index: number, elementProperties: cheerio.Element) => {
-      const imgBlog = $(elementProperties).find('img');
-      const excerptBlog = $(elementProperties).find('blockquote');
-      if (!blogImgSrc) {
-        // header image has priority over page-proterties's image
-        blogImgSrc = imgBlog?.attr('src');
-        context.setHeaderImage(blogImgSrc);
-      }
-      if (context.getExcerpt() === '') {
-        context.setExcerpt(excerptBlog.text());
-      }
-    });
+  // $(".plugin-tabmeta-details[data-macro-name='details']")
+  //   .first()
+  //   .each((_index: number, elementProperties: cheerio.Element) => {
+  //     const imgBlog = $(elementProperties).find('img');
+  //     const excerptBlog = $(elementProperties).find('blockquote');
+  //     if (!blogImgSrc) {
+  //       // header image has priority over page-proterties's image
+  //       blogImgSrc = imgBlog?.attr('src');
+  //       context.setHeaderImage(blogImgSrc);
+  //     }
+  //     if (context.getExcerpt() === '') {
+  //       context.setExcerpt(excerptBlog.text());
+  //     }
+  //   });
 
   // if not excerpt at all then alternatively we take a summary of the body of the document
   if (context.getExcerpt() === '') {