+++ src/proxy-page/proxy-page.service.ts
@@ -44,6 +44,8 @@ import addSlideTypeByStrategy from './strategySteps/addSlideTypeByStrategy';
 import addSlideContextByStrategy from './strategySteps/addSlideContextByStrategy';
 import fixCaptionImage from './steps/fixCaptionImage';
 import fixConfluenceSpace from './steps/fixConfluenceSpace';
+import addTableResponsive from './steps/addTableResponsive';
+import addAuthorVersion from './steps/addAuthorVersion';
 
 @Injectable()
 export class ProxyPageService {
@@ -90,11 +92,13 @@ export class ProxyPageService {
       spaceKey,
       pageId,
       theme,
+      type,
       style,
       content,
-      true,
+      true, // loadAsDocument
       view,
     );
+    const contextType = this.context.getType();
     const addJiraPromise = addJira(this.config, this.jira)(this.context);
     await getExcerptAndHeaderImage(this.config, this.confluence)(this.context);
     fixHtmlHead(this.config)(this.context);
@@ -117,13 +121,15 @@ export class ProxyPageService {
     fixImageSize()(this.context);
     fixCaptionImage(content)(this.context);
     fixColGroupWidth()(this.context);
-    if (type === 'blog') {
+    if (contextType.includes('blog')) {
       await addHeaderBlog()(this.context);
-    } else if (type !== 'notitle') {
+    } else if (!contextType.includes('notitle')) {
       await addHeaderTitle(this.confluence)(this.context);
     }
     fixSVG(this.config)(this.context);
     fixTableBackground()(this.context);
+    addTableResponsive()(this.context);
+    addAuthorVersion()(this.context);
     delUnnecessaryCode()(this.context);
     addCustomCss(this.config, style)(this.context);
     addLibrariesCSS()(this.context);
@@ -187,6 +193,7 @@ export class ProxyPageService {
     fixFrameAllowFullscreen()(this.context);
     fixSVG(this.config)(this.context);
     fixTableBackground()(this.context);
+    addTableResponsive()(this.context);
     delUnnecessaryCode()(this.context);
     await addJiraPromise;
     addSlideTypeByStrategy(content, this.config)(this.context);